2 containers remaining

+++ ModalPipeline/local_results/qwen3_coder_grammar_prompts_fullsuite_v1/seed_000/qwen3_coder_grammar_prompts_fullsuite_v1/eval/qwen3_coder_30b_gptq_4bit/qwen3_coder_30b_gptq_4bit_tritonbench_python_xgrammar_off_t0.0_p1.0_tok2048_one_shot_repair0_default_seed0/call_acc/exp_sqrt.py
@@ -13,7 +13,7 @@ def exp_sqrt_kernel(input_ptr, output_ptr, n_elements, BLOCK_SIZE: tl.constexpr)
 
 def exp_sqrt(input, out=None) -> torch.Tensor:
     if out is None:
-        out = torch.empty_like(input, dtype=torch.float32)
+        out = torch.empty_like(input)
     
     n_elements = input.numel()
     BLOCK_SIZE = 1024
